main_test.py
import main
import io
import sys
import re
import math
import types


def test_main_1():
    captureOut = io.StringIO()
    sys.stdout = captureOut
    # datastr = 'Kim \n 100 80 70 60\n Bill \n 100 90 80 60 \n Mary \n 90 80 70 100'
    # sys.stdin = io.StringIO(datastr)

    sys.stdout = sys.__stdout__
    print('Captured ', captureOut.getvalue())
    lines = captureOut.getvalue().split('\n')
    print(lines)

    students = main.makeStudentDictionary()
    main.printStudent(students)
    scores = main.findStudent(students, '2023-0001')

    students.sort(key=lambda d: d['Name'])

    assert len(students) == 4
    assert students[0]['Name'] == 'Bill Watson'
    assert students[3]['Name'] == 'Mary Smith'

    assert len(scores) == 3

    # regex_string = r'[\w,\W]*1'
    # regex_string += r'[\w,\W]*3'
    # regex_string += r'[\w,\W]*5'
    # regex_string += r'[\w,\W]*'
    # print(regex_string)
    # res = re.search(regex_string, main.evenlist)
    # assert res != None
    # print(res.group())
